Make it build. Requirements update and collateral.
Seems some underlying version dependency was missing from
requirements.txt and/or build-requirements.txt.  So this refreshes
dependencies based on current versions of Flask, Jinja2, et.al.

This should be reviewed with a jaded eye.  The author is neither a
front end developer nor a Python expert.  This is just a product of
pattern matching and edit, build, run cycles.